Replace RATT transformation by RocketRML, rename rattScript.ts to rocketrmlScript.ts. Still need to remove RATT script generation. Replace the white by transparent in the LDWizard logo
