+++ src/config/rmlScript.ts
@@ -32,7 +32,7 @@ async function getRmlTransformationScript(configuration: TransformationConfigura
             predicate: namedNode("csvw:url"),
             object: literal(
               usesRefinedSource
-                ? getFileBaseName(configuration.sourceFileName) + "-enriched.csv"
+                ? getFileBaseName(configuration.sourceFileName) + ".csv"
                 : configuration.sourceFileName
             ),
           },
